Changed naming scheme; Renamed utils
